requirements.txt
PyYAML==3.12
opcua==0.95.2
cryptography==2.2.2